refactor prompt chat rendering into module
